Fixed compile warning on 64bit platforms

+++ src/channel.c
@@ -17,7 +17,7 @@
  *   along with this program; if not, write to the Free Software
  *   Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
  *
- * $Id: channel.c,v 7.252 2001/07/08 03:01:06 a1kmm Exp $
+ * $Id: channel.c,v 7.253 2001/07/18 14:36:57 androsyn Exp $
  */
 #include "tools.h"
 #include "channel.h"
@@ -1268,7 +1268,7 @@ chm_simple(struct Client *client_p, struct Client *source_p,
            char **parv, int *errors, int alev, int dir, char c, void *d,
            const char *chname)
 {
-  int mode_type;
+  long mode_type;
   int i;
 
   if (alev < CHACCESS_CHANOP)
@@ -1280,9 +1280,7 @@ chm_simple(struct Client *client_p, struct Client *source_p,
     return;
   }
 
-  /* XXX this causes warnings on a 64bit compiler 
-   * sizeof(void *) > sizeof(int) */
-  mode_type = (int)d;
+  mode_type = (long)d;
 
   /* setting + */
   if (dir < 0 && !(chptr->mode.mode & mode_type))
